New Handler Working w/ Instance Of

+++ src/main/java/oogasalad/model/gameparser/data/Dimension.java
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  *
  * @author Judy He
  */
-public record Dimension(@JsonProperty("x_dimension") int xDimension,
-                        @JsonProperty("y_dimension") int yDimension) {
+public record Dimension(@JsonProperty("x_dimension") double xDimension,
+                        @JsonProperty("y_dimension") double yDimension) {
 
 }
